Clippy happy

+++ src/ranksel/sparse_index.rs
@@ -72,7 +72,7 @@ impl<B: SelectHinted, O: VSlice, const QUANTUM_LOG2: usize> Select
         let rank_at_pos = index << QUANTUM_LOG2;
 
         self.bits
-            .select_unchecked_hinted(rank, pos as usize, rank_at_pos)
+            .select_unchecked_hinted(rank, pos, rank_at_pos)
     }
 }
 
@@ -114,7 +114,7 @@ impl<B: SelectHinted + SelectZero, O: VSlice, const QUANTUM_LOG2: usize> SelectH
     #[inline(always)]
     unsafe fn select_unchecked_hinted(&self, rank: usize, pos: usize, rank_at_pos: usize) -> usize {
         let index = rank >> QUANTUM_LOG2;
-        let this_pos = self.ones.get_unchecked(index) as usize;
+        let this_pos = self.ones.get_unchecked(index);
         let this_rank_at_pos = index << QUANTUM_LOG2;
 
         // choose the best hint, as in the one with rank_at_pos closest to rank
